src/main/java/com/optimusprime/springrecipeapp/converters/UnitOfMeasureCommandToUnitOfMeasure.java
package com.optimusprime.springrecipeapp.converters;

import com.optimusprime.springrecipeapp.commands.UnitOfMeasureCommand;
import com.optimusprime.springrecipeapp.domain.UnitOfMeasure;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class UnitOfMeasureCommandToUnitOfMeasure implements Converter<UnitOfMeasureCommand,UnitOfMeasure> {

    @Synchronized
    @Nullable
    @Override
    public UnitOfMeasure convert(UnitOfMeasureCommand source) {
        if(source==null){
            return null;
        }
       final  UnitOfMeasure uom = new UnitOfMeasure();
        uom.setId(source.getId());
        uom.setDescritpion(source.getDescritpion());
        return uom;
    }
}
